fix: make sure the sources list is given a var
The `p` variable was meant to be the basis that allowed the list of sources to be viewable. I totally removed it before pushing at some point and wasn't sure why it wasn't working on campus.

It's an issue as well with the obs-rotation script, but that's to be fixed soon, as well.